change in gradient flow printing:
The content of the output file is stored in a string and only then printed on the output file.
This avoids broken gradeint flow output files in case of interrupted runs.
In this way one can stop and restart a run without having to manually correct or run the measurement on the interrupted gauge configuration.