I've added routines for the computation of loops:
It's now possible to compute at once the products of gauge links along all possible closed loops of given length.
  This is achieved through a `links::links_loops` class that is able to store in a `std::vector` all these products.
  The geometrical loops are built using an ancillary class `links::closed_paths` which deals with `links::path` objects.
I've tested the correctness of the implementation on a simple case:
       new file:   test/links.cpp

Use this new feature to implement the computation of more sophisticated glueball operators,
or to compute more than one type and use the GEVP in order to extract the energies of the lightest states.